Cross test rig for running Daffodil-flavor TDML on IBM DFDL
Dynamically loaded by Daffodil if found on the class path.

Using DFDLCatalogResolver from Daffodil to insure include/imports
resolve the same as on Daffodil.

There are test-rig-tests here, and also the sample that comes with
IBM DFDL can be run as TDML tests. You just have to copy files from the
IBM DFDL installation to the lib and src/test/resources directories.

All the tests pass on Daffodil and IBM DFDL.

See the README.txt file for how to setup and run.