Remap PUA back to XML Illegal chars on unparsing.
This is symmetric with changes that map XML-illegals to the Unicode
Private Use Area on parsing.

Ex: On parse NUL (aka U+0) becomes U+E000. On unparse U+E000 becomes 0,
that is NUL again.

This should allow daffodil TDML tests that depend on this escaping and
substitution convention to be portable to IBM DFDL as well.